requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
attrs==23.2.0
azure-cognitiveservices-speech==1.38.0
blinker==1.8.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
dashscope==1.19.3
dashvector==1.0.16
distro==1.9.0
exceptiongroup==1.2.1
Flask==3.0.3
frozenlist==1.4.1
grpcio==1.64.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
importlib_metadata==7.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
multidict==6.0.5
numpy==1.26.4
ollama==0.2.1
openai==1.34.0
protobuf==3.20.3
pydantic==2.7.4
pydantic_core==2.18.4
PyYAML==6.0.1
requests==2.32.3
sniffio==1.3.1
tqdm==4.66.4
typing_extensions==4.12.2
urllib3==2.2.1
Werkzeug==3.0.3
yarl==1.9.4
zipp==3.19.2
